# docs/swagger.yaml
basePath: /
definitions:
...
  models.Employee:
    properties:
      color:
        type: string
      created_at:
        type: string
      department_id:
        type: integer
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      last_name:
        type: string
      password:
        type: string
      updated_at:
        type: string
    type: object
  models.ShiftDay:
    properties:
      created_at:
        type: string
      date:
        type: string
      employee_id:
        type: integer
      id:
        type: integer
      notes:
        type: string
      shift_type_id:
        type: integer
      shift_week_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  responses.APIResponse:
    properties:
      data: {}
      error:
        example: Fehler bei der Verarbeitung
        type: string
      message:
        example: Operation erfolgreich
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Carsten Bröckert
  description: REST-API für die Verwaltung von Schichtplänen
  license:
    name: MIT
  title: Schichtplaner API
  version: "1.0"
paths:
...
  /api/v1/employees:
    get:
      consumes:
      - application/json
      description: Ruft alle Mitarbeiter mit ihren Berechtigungen und Schichten ab
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Employee'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Alle Mitarbeiter abrufen
      tags:
      - employees
    post:
      consumes:
      - application/json
      description: Erstellt einen neuen Mitarbeiter mit Validierungen
      parameters:
      - description: Mitarbeiter-Daten
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Employee'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Mitarbeiter erstellen
      tags:
      - employees
  /api/v1/employees/{id}:
    delete:
      consumes:
      - application/json
      description: Löscht einen Mitarbeiter
      parameters:
      - description: Mitarbeiter-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Mitarbeiter löschen
      tags:
      - employees
    get:
      consumes:
      - application/json
      description: Ruft einen spezifischen Mitarbeiter mit Details ab
      parameters:
      - description: Mitarbeiter-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Employee'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Einzelnen Mitarbeiter abrufen
      tags:
      - employees
    put:
      consumes:
      - application/json
      description: Aktualisiert einen bestehenden Mitarbeiter
      parameters:
      - description: Mitarbeiter-ID
        in: path
        name: id
        required: true
        type: integer
      - description: Aktualisierte Mitarbeiter-Daten
        in: body
        name: employee
        required: true
        schema:
          $ref: '#/definitions/models.Employee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Employee'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Mitarbeiter aktualisieren
      tags:
      - employees
  /api/v1/employees/available/{date}:
    get:
      consumes:
      - application/json
      description: Ruft alle Mitarbeiter ab, die an einem bestimmten Datum noch keine
        Schicht haben
      parameters:
      - description: Datum (YYYY-MM-DD)
        in: path
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Employee'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Verfügbare Mitarbeiter abrufen
      tags:
      - employees
  /api/v1/employees/shifts/{id}:
    get:
      consumes:
      - application/json
      description: Ruft alle Schichten eines bestimmten Mitarbeiters ab
      parameters:
      - description: Mitarbeiter-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ShiftDay'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Schichten eines Mitarbeiters abrufen
      tags:
      - employees
...
swagger: "2.0"